requirements.txt
ipaddress
Flask
paho-mqtt
pandas
numpy
statsmodels
stepwise-regression